fix bug:age>128

+++ src/main/java/com/tangcheng/db/entity/StudentDo.java
@@ -10,7 +10,7 @@ public class StudentDo {
 
     private String name;
 
-    private Byte age;
+    private Integer age;
 
     private String classes;
 
@@ -45,14 +45,14 @@ public class StudentDo {
     /**
      * @return age
      */
-    public Byte getAge() {
+    public Integer getAge() {
         return age;
     }
 
     /**
      * @param age
      */
-    public void setAge(Byte age) {
+    public void setAge(Integer age) {
         this.age = age;
     }
